+++ packages/analytics-core/src/index.ts
@@ -12,6 +12,7 @@ export { IConfig } from './types/config/core-config';
 export { Logger, ILogger, LogConfig } from './logger';
 export { getGlobalScope } from './global-scope';
 export { getAnalyticsConnector, setConnectorDeviceId, setConnectorUserId } from './analytics-connector';
+export { normalizeInstanceName } from './utils/instance-name';
 export { isNewSession } from './session';
 export { getCookieName, getOldCookieName } from './cookie-name';
 export { getLanguage } from './language';
@@ -53,6 +54,9 @@ export {
   RemoteConfigClient,
   IRemoteConfigClient,
   RemoteConfig,
+  RemoteConfigGroup,
+  RemoteConfigInfo,
+  RemoteConfigStorage,
   Source,
   RemoteConfigFetchRequest,
   RemoteConfigCustomFetch,
@@ -60,7 +64,7 @@ export {
 
 export { LogLevel } from './types/loglevel';
 export { AMPLITUDE_PREFIX, STORAGE_PREFIX } from './types/constants';
-export { Storage, IdentityStorageType, CookieStorageConfig } from './types/storage';
+export { Storage, StorageData, IdentityStorageType, CookieStorageConfig } from './types/storage';
 export { consoleObserver } from './observers/console';
 export {
   Event,
@@ -165,6 +169,7 @@ export {
   ReactNativeTrackingOptions,
   ReactNativeOptions,
   ReactNativeAttributionOptions,
+  ReactNativeStorageData,
 } from './types/config/react-native-config';
 export { ReactNativeClient, NavigationState } from './types/client/react-native-client';