+++ packages/plugin-page-view-tracking-browser/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.11.16](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-page-view-tracking-browser@2.11.15...@amplitude/plugin-page-view-tracking-browser@2.11.16) (2026-09-21)
+
+**Note:** Version bump only for package @amplitude/plugin-page-view-tracking-browser
+
+
+
+
+
+## [2.11.15](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-page-view-tracking-browser@2.11.14...@amplitude/plugin-page-view-tracking-browser@2.11.15) (2026-09-16)
+
+**Note:** Version bump only for package @amplitude/plugin-page-view-tracking-browser
+
+
+
+
+
 ## [2.11.14](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-page-view-tracking-browser@2.11.13...@amplitude/plugin-page-view-tracking-browser@2.11.14) (2026-08-24)
 
 **Note:** Version bump only for package @amplitude/plugin-page-view-tracking-browser